vk_bot/requirements.txt
aiohttp==3.8.5
aiohttp-apispec==2.2.3
aiohttp-session==2.12.0
aiosignal==1.3.1
apispec==3.3.2
async-timeout==4.0.3
asyncpg==0.28.0
attrs==23.1.0
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.4
frozenlist==1.4.0
greenlet==2.0.2
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.3
marshmallow==3.20.1
multidict==6.0.4
packaging==23.1
pycparser==2.21
PyYAML==6.0.1
SQLAlchemy==2.0.21
typing_extensions==4.8.0
webargs==5.5.3
yarl==1.9.2
alembic==1.12.0
